+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+### Added
+- `react/jsx-indent` new rule from eslint-plugin-react v3.14.0
+- pragma specified for eslint-plugin-react
 
 
 ## [3.2.0] - 2015-12-26